[VW-451] Moving Agent code into src/features/agent (#223)
* [VW-451] Updated CLAUDE.md 

* [VW-451] Moving Agent code into a top level agents directory

* [VW-451] Rabbit suggestions
