+++ src/features/agents/shared/history.ts
