+++ src/features/agents/shared/notes-preload.ts
