+++ src/features/agents/shared/stream-bridge.ts
