+++ src/features/agents/tools/query-platform-tool.ts
